+++ src/metar/models/cloud.rs
@@ -3,7 +3,7 @@
 //! Contains types and parsing logic implemented for this crate.
 use serde::Serialize;
 
-#[derive(Debug, Serialize, PartialEq, Eq)]
+#[derive(Debug, Serialize, PartialEq, Eq, Clone)]
 /// Defines the CloudLayer domain model used by the parser.
 pub struct CloudLayer {
     pub amount: CloudAmount,
@@ -11,7 +11,7 @@ pub struct CloudLayer {
     pub cloud_type: Option<CloudType>,
 }
 
-#[derive(Debug, PartialEq, Eq, Serialize)]
+#[derive(Debug, PartialEq, Eq, Serialize, Clone)]
 /// Enumerates the allowed values for CloudAmount.
 pub enum CloudAmount {
     FEW,
@@ -23,7 +23,7 @@ pub enum CloudAmount {
     VV,
 }
 
-#[derive(Debug, PartialEq, Eq, Serialize)]
+#[derive(Debug, PartialEq, Eq, Serialize, Clone)]
 /// Enumerates the allowed values for CloudType.
 pub enum CloudType {
     CB,